requirements.txt
quart
quart-cors
python-dotenv
mistralai
werkzeug
openai